+++ changelog.md
@@ -1,5 +1,18 @@
 # Unreleased
 
+# 3.0.1
+
+- Add SaberFaction and FactionUUID support
+- Add command ``/koth addcommand <koth name> <start/win> <command>``
+- Add comment in file koth-example.yml
+- Add placeholders:
+1. ``%zkoth_team_id%`` - Current koth capture player team id
+2. ``%zkoth_team_leader%`` - Current koth capture player team leader name
+3. ``%zkoth_active_<koth name>%`` - Lets you know if a koth is active
+4. ``%zkoth_cooldown_<koth name>%`` - Lets you know if a koth is in cooldown
+5. ``%zkoth_start_<koth name>%`` - Lets you know if a koth is start
+- Fix scoreboard and hologram update
+
 # 3.0.0
 
 New version completely recoded